image_processing/requirements.txt
fastapi==0.73.0
hypercorn==0.13.2
python-multipart==0.0.5
Pillow==9.0.0
google-cloud-storage==2.1.0
google-cloud-datastore==2.4.0